Added form automation.
This commit greatly extends the functionality by supporting every state that a real form checkbox can, and improving the synchronization between checkboxes in a form and the toggles. We add a "disabled" state (50% opacity), which prevents the toggle from being clicked   or dragged until the state is lifted.

Script based modification of the checkbox's state with jQuery.prop() will now cause the toggle's state to be updated. In the same way, modifying the state of the toggle will update the checkbox.